requirements.txt
Django==3.0.11
djangorestframework==3.11.2
django-filter==2.4.0
